[IMM32] Restructure! (Retry)
